fix: dont reread files in sans2d workflow
